separated TestPredict class definition to a new header file so that it can be reused across test files
